Version 1.0.0 ready to go
Finalizes lit_update.py, mendeley_to_org.py, README (now .org instead
of .md). Adds lit_add.py, litearture.el, pybib_utils.py

mendeley_to_org.py:

 - Adds some comments

 - Moved paper_dir and org_format to lit_add.py, loads from there

 - Raises exception at beginning if master bib or org file found

 - Adds few extra print statements

 - Goes through the org and bib statements in a sorted manner, so the
   resulting master files are sorted

lit_update.py:

 - modified_time now separate for org and bib file, local pdf and org
   files checked against master org, local bib files checked against
   master bib.

 - Only checks pdf file if there's one in the bibtex entry and it
   exists

 - Added assertions to make sure that substiutions are made when I
   think they are

 - Added updates for bib and org files.

lit_add.py: contains functions to add new pdf or bib file to the
library. If it's a pdf, attempts to find bibtex entry. If it's a
bibtex entry, checks file field; if the field exists and the file
exists, treats them as paired, else assumes there's no file for this
entry. Creates new directory, org note, and moves everything within
the directory. Adds new information to master bibtex and org files.

pybib_utils.py: Snippets of code from the pybib module, used to get
bibtex entry from a doi from within python.

literature.el: code used to make emacs play nice with my setup. Adds
helm, helm-bibtex integration.x

+++ lit_update.py
@@ -7,38 +7,14 @@
 #  - If notes updated: update master notes
 #  - If bib updated: update master bib
 
-# Decide if I want to make the master bib and org files unable for me
-# to edit or try and find a way to merge any edits. There's a python
-# difflib that would help with this. -- I think for now, make the
-# "master" files un-writable, and just update them to reflect the
-# changes in the sub-files.
-
-#Also need a function for adding a file to library. That should
-#intialize the note file (move org_format to new script, have
-#mendeley_to_org grab it from there), make directory, rename pdf and
-#move it there. Require bib to be downloaded as well or get that
-#automatically? Hard without doi, and getting a doi automatically not
-#guaranteed.  Can try to get doi like so:
-"""
-regex=re.compile('\\b(10[.][0-9]{4,}(?:[.][0-9]+)*/(?:(?![\"&\'<>])\S)+)\\b')
-
-doc = poppler.document_new_from_file(path,None)
-pages = [doc.get_page(i) for i in range(doc.get_n_pages())]
-doi = []
-for page in pages:
-    text = page.get_text()
-    doi += re.findall(regex,text)
-#check they're all the same and peace out
-"""
-#If got doi, then just do the following. Else, give option to enter doi, then tell user to manually download bib.
-
-#With pybib installed, can just run `bib DOI` from the command line and get its bib file. Can run from within python though?
-#Actually, just use the code that underlies it, found at https://github.com/jgilchrist/pybib/blob/master/pybib/utils.py
-
+#This file only updates one-way: from local files to master files. For
+#this reason, the master files are kept read-only. If this becomes too
+#annoying, I may try to find a way to update both ways, but for now
+#this works.
 
 def update(fill_column=70):
     import os,stat,bibtexparser,time,re
-    from mendeley_to_org import paper_dir,org_format
+    from lit_add import paper_dir,org_format
     
     #Make it so we can write again...
     os.chmod(paper_dir+'/literature.bib',stat.S_IWUSR|stat.S_IREAD)
@@ -48,28 +24,53 @@ def update(fill_column=70):
         paper_dir=paper_dir[:-1]
     
     with open(paper_dir+'/literature.bib') as f:
-        bib_db = bibtexparser.loads(f.read())
+        bib_db = bibtexparser.load(f)
 
-    modified_time = os.path.getmtime(paper_dir+'/literature.bib')
-    assert modified_time==os.path.getmtime(paper_dir+'/literature.bib'),'Master bib and org files edited separately somehow!'
+    modified_time = {'bib':os.path.getmtime(paper_dir+'/literature.bib'),'org':os.path.getmtime(paper_dir+'/literature.org')}
     for bib in bib_db.entries:
-        dir_path = os.path.expanduser(os.path.split(bib['file'])[0])
         bib_id = bib['ID']
-        if os.path.getmtime('%s/%s.pdf'%(dir_path,bib_id)) > modified_time:
+        if 'file' in bib: 
+            dir_path = os.path.expanduser(os.path.split(bib['file'])[0])
+        else:             
+            dir_path = paper_dir+'/%s'%bib_id
+        if os.path.isfile('%s/%s.pdf'%(dir_path,bib_id)) and os.path.getmtime('%s/%s.pdf'%(dir_path,bib_id)) > modified_time['org']:
+            print('Pdf %s.pdf updated, extracting annotations'%bib_id)
             annots = get_annotations('%s/%s.pdf'%(dir_path,bib_id),'org',org_indent=re.search('Annotations *\n( *){annotations}',org_format).groups()[0],fill_column=fill_column)
             with open('%s/%s.org'%(dir_path,bib_id),'r+') as f:
                 org_note = f.read()
-                org_note = re.sub("Annotations.*\n([*]+)","Annotations\n\n%s\n\n%s"%(annots,r"\1"),org_note,flags=re.DOTALL)
+                re_sub = re.subn("Annotations.*\n([*]+)","Annotations\n\n%s\n\n%s"%(annots,r"\1"),org_note,flags=re.DOTALL)
+                assert re_sub[1]==1,'Thought pdf was updated but no (or more than one...) substitution made...'
                 f.seek(0)
-                f.write(org_note)
+                f.write(re_sub[0])
                 f.truncate()
-        if os.path.getmtime('%s/%s.org'%(dir_path,bib_id)) > modified_time:
-            #Replace section in master org with new local org (use regex?)
-            'hi'
-        if os.path.getmtime('%s/%s.bib'%(dir_path,bib_id)) > modified_time:
-            #Replace section in master bib with new local bib (use regex?)
-            'bye'
-            
+        if os.path.getmtime('%s/%s.org'%(dir_path,bib_id)) > modified_time['org']:
+            print('Org file %s.org updated, copying new changes to master org'%bib_id)
+            with open('%s/%s.org'%(dir_path,bib_id)) as f:
+                org_note = f.read()
+            #Drop everything before the first header (startup options, etc)
+            org_note = org_note[org_note.find('*'):]
+            with open(paper_dir+'/literature.org','r+') as f:
+                master_org = f.read()
+                #Make use of the fact that we know the bib key to find
+                #and replace this specific entry. The |$ is because it
+                #could be the last entry, in which case there's not
+                #another one after it.
+                re_sub = re.subn('(.*)[*] [A-Z]* .* \t+.*:BIBTEX-KEY: %s.*?(\n[*] [A-Z]*|$)'%bib_id,r"\1"+org_note+'\n'+r"\2",master_org,flags=re.DOTALL)
+                assert re_sub[1]==1,'Thought org file was updated but no (or more than one) substitution made...'
+                f.seek(0)
+                f.write(re_sub[0])
+                f.truncate()
+        if os.path.getmtime('%s/%s.bib'%(dir_path,bib_id)) > modified_time['bib']:
+            print('Bib file %s.bib updated, copying new changes to master bib'%bib_id)
+            with open('%s/%s.bib'%(dir_path,bib_id)) as f:
+                bib_file = f.read()
+            with open(paper_dir+'/literature.bib','r+') as f:
+                master_bib = f.read()
+                re_sub = re.subn('@[a-zA-z]*{%s,.*(@|$)'%bib_id,bib_file+r"\1",master_bib,flags=re.DOTALL)
+                assert re_sub[1]==1,'Thought bib file was updated but no (or more than one) substitution made...'
+                f.seek(0)
+                f.write(re_sub[0])
+                f.truncate()
             
     #We want these files to be read-only so that the only edits are to
     #the individual org and bib files. If this is too inconvenient,
@@ -135,3 +136,5 @@ def col_wrap(text,fill_col,org_indent=''):
     text = ' '.join(text).replace("\n ","\n")
     return text
     
+if __name__ == '__main__':
+    update()
